If the shape isn't stored in basic flags, set the size pool to 1
On 32 bit machines, we can't store the shape in basic object flags.
On 32 bit machines, setting a shape that corresponds to the size pool
would be very onerous, so we're _enabling_ RVARGC on 32 but machines,
but decreasing the size pool count to 1.  That way there is only one
"size pool / capacity" shape.